# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocSetu Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: gender
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkerProfile"
...
  /workers/me:
    get:
      operationId: getMyWorkerProfile
      tags: [workers]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: My worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProfile"
    post:
      operationId: createWorkerProfile
      tags: [workers]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProfile"
    patch:
      operationId: updateWorkerProfile
      tags: [workers]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProfile"
...
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkerProfile:
      type: object
      required: [id, userId, name, skills, location, rating, reviewCount, isAvailable, isVerified, completedJobs, languages]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        isVerified:
          type: boolean
        completedJobs:
          type: integer
        experience:
          type: ["string", "null"]
        languages:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        hourlyRate:
          type: ["number", "null"]
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    WorkerProfileInput:
      type: object
      required: [skills, location]
      properties:
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experience:
          type: string
        bio:
          type: string
        hourlyRate:
          type: number
        languages:
          type: array
          items:
            type: string
    WorkerProfileUpdate:
      type: object
      properties:
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        experience:
          type: string
        bio:
          type: string
        hourlyRate:
          type: number
        isAvailable:
          type: boolean
        languages:
          type: array
          items:
            type: string
...